Extract class refactoring
